Added LocationUtils + changed few smaller things
