+++ src/main/java/me/partlysanestudios/partlysaneskies/dungeons/RequiredSecretsFound.java
@@ -8,6 +8,7 @@ package me.partlysanestudios.partlysaneskies.dungeons;
 import me.partlysanestudios.partlysaneskies.PartlySaneSkies;
 import me.partlysanestudios.partlysaneskies.system.BannerRenderer;
 import me.partlysanestudios.partlysaneskies.system.PSSBanner;
+import me.partlysanestudios.partlysaneskies.utils.LocationUtils;
 import me.partlysanestudios.partlysaneskies.utils.Utils;
 import me.partlysanestudios.partlysaneskies.utils.TabListUtils;
 import net.minecraft.client.Minecraft;
@@ -24,11 +25,11 @@ public class RequiredSecretsFound {
 
     @SubscribeEvent
     public void checkRequiredSecrets(TickEvent.ClientTickEvent event) {
-        if (!PartlySaneSkies.isSkyblock()) {
+        if (!LocationUtils.isSkyblock()) {
             return;
         }
 
-        if (!Utils.inDungeons()){
+        if (!LocationUtils.inDungeons()){
             return;
         }
